moved lexer utility functions to separate file
